move WakeCheckFilter to processors/filters
